Created default_tile.json
Will hold all the data for the tiles in the default deck, naming conventions are [Cloyster/Castle/Road/Grass][Directions piece is present][order-number].

At the moment I was planning on using a 3x3 grid to represent placeable tiles and tile types, however based on the complexities of a few pieces I may have to upscale to 5x5. Not complete atm either.